+++ static/js/server_events_dispatch.js
@@ -263,11 +263,11 @@ export function dispatch_normal_event(event) {
         case "realm_bot":
             if (event.op === "add") {
                 bot_data.add(event.bot);
-            } else if (event.op === "remove") {
+            } else if (event.op === "deactivate") {
                 bot_data.deactivate(event.bot.user_id);
                 event.bot.is_active = false;
-            } else if (event.op === "delete") {
-                blueslip.info("ignoring bot deletion for live UI update");
+            } else if (event.op === "change_owner") {
+                blueslip.info("ignoring bot deletion due to change_owner for live UI update");
                 break;
             } else if (event.op === "update") {
                 bot_data.update(event.bot.user_id, event.bot);