+++ zerver/lib/events.py
@@ -630,13 +630,13 @@ def apply_event(
         if event["op"] == "add":
             state["realm_bots"].append(event["bot"])
 
-        if event["op"] == "remove":
+        if event["op"] == "deactivate":
             user_id = event["bot"]["user_id"]
             for bot in state["realm_bots"]:
                 if bot["user_id"] == user_id:
                     bot["is_active"] = False
 
-        if event["op"] == "delete":
+        if event["op"] == "change_owner":
             state["realm_bots"] = [
                 item for item in state["realm_bots"] if item["user_id"] != event["bot"]["user_id"]
             ]