+++ templates/zerver/api/changelog.md
@@ -10,6 +10,11 @@ below features are supported.
 
 ## Changes in Zulip 4.0
 
+**Feature level 44**
+
+* The `remove event` for `realm_bot` is renamed to `deactivate`.
+* The `delete` event for `realm_bot` is renamed to `change_owner`.
+
 **Feature level 43**
 
 * [`GET /users/{user_id_or_email}/presence`]: Added support for